+++ TAO/examples/POA/On_Demand_Loading/Servant_Manager.cpp
@@ -81,7 +81,7 @@ ServantManager_i::obtain_servant (const char *str,
   // Cannot go from void* to function pointer directly. Cast the void*
   // to long first.
   void *symbol = dll->symbol (create_symbol_.c_str ());
-  intptr_t function = reinterpret_cast<intptr_t> (symbol);
+  long function = reinterpret_cast<long> (symbol);
 
   SERVANT_FACTORY servant_creator =
     reinterpret_cast<SERVANT_FACTORY> (function);
